Chapter 5, Exercise 13: tail

+++ chapter5/e1.c
@@ -32,5 +32,5 @@ int getint(int *pn)
     *pn *= sign;
     if (c1 != EOF)
         ungetch(c1);
-    return c1;
+    return (size_t) c1;
 }
